requirements.txt
Leonardo-Ai-SDK==3.2.4
openai==1.8.0
python-dotenv==1.0.0